projeto-leads-marketing/pyproject.toml
[project]
name = "projeto-leads-marketing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apache-airflow>=3.1.7",
    "gspread>=6.2.1",
    "oauth2client>=4.1.3",
    "pandas>=3.0.1",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.47",
]
